% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wb.R
\name{create_excel_wb}
\alias{create_excel_wb}
\title{Create an excel workbook}
\usage{
create_excel_wb(pdInput, psInFile, psOutFile, psSheetName, psTableStyle,
  pnHeaderTextRotation, pnStartRow, pnStartCol)
}
\arguments{
\item{pdInput}{input data frame}

\item{psInFile}{name of the input file}

\item{psOutFile}{name of the output file}

\item{psSheetName}{name of the sheet to be used}

\item{psTableStyle}{pre-defined table style to be used}

\item{pnHeaderTextRotation}{rotation angle for header text}
}
\description{
\code{create_excel_wb} creates an excel workbook from an R dataframe
}
\details{
The only input that must be specified is the dataframe to be converted or
a csv file that can be read into a dataframe. From that an excel workbook
is created using some standard format.
}
\examples{
create_excel_wb(pdInput = iris)

}

