% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RZwToExcel-package.r
\docType{package}
\name{RZwToExcel}
\alias{RZwToExcel}
\alias{RZwToExcel-package}
\title{RZwToExcel Package}
\description{
This package takes a dataframe and writes it to an Excel workbook.
}
\details{
The abreviation "Zw" in the package name stands for the German
word "Zuchtwert" which stands for breeding value. This package
originated in a task that a colleague got results from a run
of predicting breeding values for beef cattle.

Clients who want to work with the predicted breeding values require
to have the breeding values in an Excel workbook.

This package provides a small set of functionlities to export
data from a dataframe in R into an Excel workbook
}

